增加并且更改3d识别

+++ ThrPoint.cs
@@ -8,18 +8,18 @@ namespace KinectGetData
 {
     public struct ThrPoint
     {
-        public float X;
-        public float Y;
-        public float Z;
+        public double X;
+        public double Y;
+        public double Z;
         /// <summary>
         /// 三维结构体的赋值
         /// </summary>
         /// <param name="X"></param>
         /// <param name="Y"></param>
         /// <param name="Z"></param>
-        public ThrPoint(float X, float Y, float Z)
+        public ThrPoint(double X, double Y, double Z)
         {
-            if (X == null || Y == null || Z == null) throw new ArgumentException();
+            if ( X == null || Y == null || Z == null) throw new ArgumentException();
             this.X = X;
             this.Y = Y;
             this.Z = Z;
